chore(classical_register): move ClassicalRegister class in the utils directory
